+++ indico/modules/events/registration/client/js/form/fields/SingleChoiceInput.jsx
@@ -13,6 +13,7 @@ import {Field} from 'react-final-form';
 import {useSelector} from 'react-redux';
 import {Form, Label, Dropdown} from 'semantic-ui-react';
 
+import Combobox from 'indico/react/components/Combobox';
 import {FinalCheckbox, FinalDropdown, FinalField, parsers as p} from 'indico/react/forms';
 import {Translate} from 'indico/react/i18n';
 
@@ -75,53 +76,43 @@ function SingleChoiceDropdown({
     );
   }
 
-  const options = choices.map(c => ({
-    key: c.id,
-    value: c.id,
-    disabled:
-      !c.isEnabled ||
-      isPaidChoiceLocked(c) ||
-      (c.placesLimit > 0 && (placesUsed[c.id] || 0) >= c.placesLimit && !existingValue[c.id]),
-    text: c.caption,
-    content: (
-      <div styleName="dropdown-text">
-        <div styleName="caption">
-          <ChoiceLabel choice={c} management={management} paid={isPaidChoice(c)} />
-        </div>
-        <div styleName="labels">
-          {!!c.price && (
-            <Label>
-              {c.price} {currency}
-            </Label>
-          )}
-          {c.placesLimit === 0 ? null : (
-            <PlacesLeft
-              placesLimit={c.placesLimit}
-              placesUsed={placesUsed[c.id] || 0}
-              isEnabled={c.isEnabled}
-            />
-          )}
-        </div>
+  const options = choices.map(c => [
+    c.caption,
+    <div styleName="dropdown-text" hidden={c.disabled} key={c.id}>
+      <div styleName="caption">
+        <ChoiceLabel choice={c} management={management} paid={isPaidChoice(c)} />
       </div>
-    ),
-  }));
+      <div styleName="labels">
+        {!!c.price && (
+          <Label>
+            {c.price} {currency}
+          </Label>
+        )}
+        {c.placesLimit === 0 ? null : (
+          <PlacesLeft
+            placesLimit={c.placesLimit}
+            placesUsed={placesUsed[c.id] || 0}
+            isEnabled={c.isEnabled}
+          />
+        )}
+      </div>
+    </div>,
+  ]);
+
+  const handleChange = evt => {
+    const selected = choices.find(c => c.caption === evt.target.value);
+    onChange(selected ? {[selected.id]: 1} : {});
+  };
 
   return (
     <Form.Group styleName="single-choice-dropdown">
       <Form.Field>
-        <Dropdown
+        <Combobox
           id={id}
-          selection
-          style={{width: 500, opacity: 1}}
-          placeholder={Translate.string('Choose an option')}
+          onChange={handleChange}
           options={options}
-          disabled={disabled}
-          value={(!isPurged && selectedChoice.id) || ''}
-          onChange={(e, data) => onChange(data.value ? {[data.value]: 1} : {})}
-          clearable={!isRequired}
-          search
-          selectOnNavigation={false}
-          selectOnBlur={false}
+          value={(!isPurged && selectedChoice.caption) || ''}
+          required={isRequired}
         />
       </Form.Field>
       {extraSlotsDropdown}