+++ CHANGES.rst
@@ -67,6 +67,7 @@ Accessibility
 ^^^^^^^^^^^^^
 
 - Improve registration form date picker accessibility (:pr:`6371`, thanks :user:`foxbunny`)
+- Improve registration form single choice input accessibility (:pr:`6310`, thanks :user:`foxbunny`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^
@@ -74,6 +75,7 @@ Internal Changes
 - Use unguessable URLs for user avatar pictures (:pr:`6346`, thanks :user:`vtran99`)
 - Add ``<ind-date-picker>`` custom element (:pr:`6371`, thanks :user:`foxbunny`)
 - Use native ESM for webpack config files (:pr:`6389`)
+- Add ``<ind-combobox>`` custom element (:pr:`6310`, thanks :user:`foxbunny`)
 
 
 Version 3.3.2